browser-support: remove access to /var/lib/snapd/desktop/applications
In thinking through why /var/lib/snapd/desktop/applications is needed
(for not regressing unity messaging menu), the snap necessarily must
plugs desktop-legacy and/or unity7 in order to have the dbus accesses.
Snaps that currently plugs browser-support with allow-sandbox: true
cannot regress this functionality because they already do not have it.
As such, remove the accesses to /var/lib/snapd/desktop/applications.