boot/many: add earlyBootState and use with InitramfsRunModeSnapsToMount
This adds InitramfsRunModeSnapsToMount, which is to be used from snap-bootstrap
to move the logic about updating boot state during that stage to live alongside
the running boot state we do in normal userspace after the pivot_root/initramfs.

This implements earlyBootState with a single method, chooseSnapInitramfsMount(),
which takes a bootStateUpdate and returns a bootStateUpdate so that one can
chain together multiple snap mount choices and commit them at once at the end.
Currently that just amounts to writing the modeenv, but could be more
complicated if we needed it to be.

This also implements earlyBootState for UC20 for the snapd snap using the seed
package and duplicating the helper from cmd_initramfs_mounts.go to only load
certain snaps from the seed during the initramfs.

Still to do is to implement the install mode mounts to further consolidate with
some of the still duplicated code in cmd_initramfs_mounts.go.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>