boot,many: add modeenv.WriteTo, make Write take no args
This is helpful in situations where we don't want to carry around a directory of
where to write the modeenv when we are mutating it.

Also make some changes of s.runmnt to dirs.RunMnt in preparation for another PR
which eliminates s.runmnt in favor of using dirs.RunMnt everywhere. Making this
change here eliminates the likelihood of merge conflicts with that followup PR.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>