tests: cleanup security-private-tmp properly
The "snap-confine-undesired-mode-group" test sometimes fails. It
turns out this is because the "security-private-tmp" test leaves
a file that is not root owned in /tmp.

The snap-confine-undesired-mode-group is arguable testing too
much/the wrong thing. But the "security-private-tmp" is also
not properly cleaning up after itself so this commit fixes this
missing cleanup.