cmd/snap: handle ChildExited erorr
When invoking a snap as a non-root user talk to the session bus
to create systemd transient scope. In the case that the session bus
DBus socket is not available, such as on Debian derivatives which package
that socket into a distinct package, we get a ChildExited error when
attempting to send a DBus message.

Since the session bus is always available in desktop sessions this
failure is deemed non-fatal. It is more important to start applications
than to track them for refresh-app-awareness.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>