cmd/snap-bootstrap/initramfs-mounts: verify the seed snap matches bootloader env
The recovery system has a boot env which contains what specific kernel should be
run/mounted in the snapd_recovery_kernel env var and this is supposedly what
kernel was booted to get us to install mode, but we should verify this matches.

Also add unit tests for the happy path and common unhappy paths.

Also delete the cross-check TODO, as recoverySystemEssentialSnaps already is
verifying the snap's signatures are valid and this is probably sufficient checks
for install mode.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>