sandbox/cgroup: add ProcessPathInTrackingCgroup
This function can be used to determine the hierarchy of a given
process in the tracking cgroup used by systemd. Systemd uses either
the unified cgroup, if available (either in pure or hybrid mode) or the
special controller-less name=system cgroup.

This code will be used by the refresh app awareness logic shortly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>