tests: make session tool way more robust
This patch greatly improves the robustness of session-tool.  After
debugging the failure we've seen for several weeks I got to the smoking
gun: in the failing case, the session that was being started by session
tool was interrupted by asynchronous shutdown of the session started
just a moment before. This is because the test was using session-tool to
run several commands, one after another.

    Failed to create session: Start job for unit user-12345.slice failed with 'canceled'

This has lead to the failure in pam_systemd.so, which effectively
resulted in failure to spawn a session in logind, which resulted in no
an error message, rather than the message that was carrying the value of
XDG_RUNTIME_DIR (returned by logind in response to a request from
pam_systemd.so).

There are two new options to `session-tool`, namely `--prepare` and
`--restore`, both of which also take `-u USER` as an option (defaulting
to root if omitted). Those two options can be used to allow a persistent
session-tool to provide a session for the given user. Internally this
enables lingering, as described by loginctl(1), which means the session
will no longer be constructed and torn down by each execution of
session-tool, avoiding the entire race problem.

In addition, --prepare and --restore work around bugs in systemd-logind,
that prevent lingering from really working. In addition, on core-16
systems where /var/lib/systemd is not writable, /var/lib/systemd/linger
is constructed, for the duration of the test, using a writable mimic.

The test exercising this script has been expanded to run several hundred
iterations, in order to give more confidence that the result is no
longer racy. The test also checks both running sessions as "test" user,
as before and as the root user.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>