osutil: make the TestWriter() test less racy
We sometimes see failures in the osutil TestWriter() test. The issue
is that it gives io.Copy() 1/100s time to copy any data. On a very
busy machine that is not enough and no data is copied.

This commit increases the time limit to 1/10s which is still fast
enough to not slow down the tests but decreases the risk of hitting
this by at least ~10x.

LP: 1869889