cmd/snap: verify effective path in the tracking cgroup
When running under systemd << v238 we may end up asking systemd to
start a transient scope with our PID in it, get no notification of error
of any kind but get a silent error.

This is caused by the fact that systemd relies on kernel policy when
moving a process within a cgroup hierarchy. In the case that the same
user owns the origin and destination, the move is allowed.

In the case that this ownership is not the same this move is denied.

Systemd v238 has grown extra features where systemd-wide systemd can
assist in this move, even if the kernel would otherwise not allow it
(root can move anything). On system running older versions of systemd
this operation can silently fail.

It is important to know that it doesn't fail in a desktop session,
because ownership is arranged such so that the running user owns both
"sides".

This particular error can happen when snaps are started having logged
in remotely, for example with ssh.

To detect this case, where eventually we would like to issue a snapd
warning, measure the effective path in the tracking cgroup and log
a debug message when this situation has occurred.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>