boot/bootstate20: always cleanup try-kernel.efi and modeenv in markSuccessful
When we are marking a boot successful, we should always cleanup the modeenv and
any try-kernels on the bootloader because we currently have numerous bugs
whereby we don't effectively cleanup parts of the modeenv or try-kernel symlinks
when we fail and have to rollback.

Also, when we call boot.MarkBootSuccessful, internally genericMarkSuccessful
should always return some snap, because we did boot something, so that then in
commit() we will mark a snap as successful. This is the key part to cleaning up
properly.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>