usersession: extend timerange in TestExitOnIdle
The TestExitOnIdle test checks that the idle time is extended
when making a second request. It checks both a lower and a upper
bound but the upper bound is too low for slow (overcommitted)
machines and it fails sometimes during a PPA build. This commit
increases the uppper limit. This should be fine because this
is mostly about not exiting too early.