interfaces: don't use the owner modifier for files shared via document portal
The document portal does not alter the stat information of the files it
shares.  For example, if a user tries to open some documentation in
/usr/share/doc via a portal file chooser, the proxied file under
$XDG_RUNTIME_DIR/doc will be owned by root.

As the document portal is a FUSE filesystem, there are kernel level
restrictions to prevent other users accessing the file system (even
root).  So this shouldn't result in any effective change in access.