tests: skip "/etc/machine-id" in "writablepaths" test
When systemd is booting it's playing tricks with /etc/machine-id
on firstboot. I.e. it is creating a RO tmpfs on top of the real
machine-id file:
```
google:ubuntu-core-20-64 .../tests/core/writablepaths# findmnt /etc/machine-id
TARGET        SOURCE                                             FSTYPE OPTIONS
/etc/machine-id
              /dev/mapper/ubuntu-data[/system-data/etc/machine-id]
                                                                 ext4   rw,rela
/etc/machine-id
              tmpfs[/machine-id]                                 tmpfs  ro,size

```
This causes the test to fail for this particular file. This commit
skips the test for /etc/machine-id.