packaging: use debian/not-installed to ignore snap-preseed
The debian build failed because snap-preseed was not removed
in the "_dh_install-indep" target. This made it fail during
the amd64 build on Debian. Instead of removing the file in
both the "_dh_install-{arch,indep}" target this commit uses
debian/not-installed to avoid the duplication.

This should have been caught in the nightly sbuild-* test
but the test fails there since forever due to an incorrect
setup.