github: give jobs shorter names
GitHub actions constructs a line out of workflow name, job name and
matrix configuration. At some point the line length limit is exceeded
and we loose the tail of the line, which holds the duration of the run.

In order to get the duration to be visible, use shorter terms.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>