osutil: use exported ProcSelfMountInfo var in real funcs
This was just using an internal unexported variable so we could modify that for
tests, but now ProcSelfMountInfo is prefixed with dirs.GlobalRootDir, so we
don't need to mock that and can instead just use the exported variable directly.

Also add a helper for tests across the codebase to be able to write out a
ProcSelfMountInfo file if needed to remove that duplicated snippet of
os.MkdirAll and ioutil.WriteFile.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>