template/network: updates for libnss-systemd
allow accessing the libnss-systemd VarLink sockets and DBus APIs

https://systemd.io/USER_GROUP_API/ describes the libnss-systemd VarLink
socket APIs:

"
When a client wants to look up a user or group record, it contacts all
sockets bound in this directory in parallel, and enqueues the same query
to each. The first positive reply is then returned to the application,
or if all fail the last seen error is returned instead. (Alternatively
a special Varlink service is available, io.systemd.Multiplexer which
acts as frontend and will do the parallel queries on behalf of the
client, drastically simplifying client development.)
"

This updates the default template to allow read/write on well-known
systemd VarLink named sockets.

In addition, update the network interface to allow allow lookups for
systemd-exec's DynamicUsers via D-Bus (since it already has access to
DBus for resolved.

References:
- https://systemd.io/USER_GROUP_API/
- https://systemd.io/USER_RECORD/
- https://www.freedesktop.org/software/systemd/man/nss-systemd.html
- https://www.freedesktop.org/software/systemd/man/systemd.exec.html
- https://launchpad.net/bugs/1796911
- https://launchpad.net/bugs/1869024
- https://gitlab.com/apparmor/apparmor/-/merge_requests/474