many: osutil should not depend on dirs, mv ProcSelfMountInfo to dirs
We now have ProcSelfMountInfo defined inside dirs instead of osutil, so that it
can be updated when rootdir changes without needing osutil to import dirs.

This does mean that some of the helpers from osutil now unfortunately take an
argument, which is where to read the mountinfo from.

Also eliminate yet another helper for mocking mountinfo in gadget, and simplify
selinux mountinfo mocking to use the osutil helper.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>