osutil: don't export ProcSelfMountInfo, use mocked string
Many tests currently implicitly rely on reading state from /proc/self/mountinfo
through the system-key or other various aspects that care about what's mounted
on the system. These tests currently suffer from leaking state from the host
system that the unit test is being run on into the test, exposing us to a
problem where we may have false positives if these tests are run on systems with
the state present and the state was not properly mocked.

To alleviate this, we make LoadMountInfo take no arguments and panic if it was
not properly mocked during tests. The proper way to mock mountinfo now is to use
MockMountInfo(), which is now exported and will do the right thing.

This does mean we have to eliminate a test in osutil which was testing about the
filename argument to LoadMountInfo which no longer is possible.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>