github: cancel previous builds
GitHub actions does not have a first party feature of cancelling
workflows that are CI related when a new commit is pushed to an open
pull request.

This puts a lot of load on our queue that is processing effectively
stale requests. To ease the pain integrate with a 3rd party action
to see if we can get the desired effect.

If successful the action will be forked and hosted on snapcore/*, to
ensure we are in control.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>