travis.yml: re-enable travis
Moving to GH actions was a worthwhile try but ultimately it was
premature. While GH actions are great in pretty much every
aspect, the following missing feature is a show-stopper IMO:

The most important missing feature for GH actions is that it
does not provide a way to automatically cancel jobs that
are no longer relevant. Like when a new revision is pushed to
a PR that was not tested yet there will be a new action created
but the existing action will also run. This creates huge amounts
of jobs in GCE and the wait time is overly long. Once this missing
auto-cancel feature is available we can revisit GH actions again.