boot/efi_amd64.go: fallback to sysfs for EFI vars, prefer efivarfs
There are 2 possible filesystems that we could read EFI variables from, the old,
deprecated sysfs interface at /sys/firmware/efi/vars, which requires reading the
value with a "data" file, and the new efivarfs filesystem which is typically
mounted at /sys/firmware/efivars. We should prefer the new efivarfs interface,
as it more fully supports the final EFI standards (including things like EFI
var data that is longer than 1024 bytes). However, if we don't have that, or if
we can't find it because there is something wrong with procfs and
/proc/self/mountinfo, then we should fallback on the sysfs interface. In
general, this code should be maximally robust to protect against bricking a
device, since this will be used in the initramfs with snap-bootstrap.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>