cmd/snap-bootstrap/partition/lsblk: add Equals method
This will allow us to cross-check that partitions mounted at different mount
points come from the same disk.

Also add MajorMinor to the lsblk output and use this to compare *disk structs.
MajorMinor will be the same for the actual block device for a *disk, note we are
not comparing partition block devices, but the actual disk block device here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>