cmd/snap-bootstrap/initramfs-mounts: use ubuntu-boot to find ubuntu-data{,-enc}
Before we try to decrypt ubuntu-data-enc, we should locate it by searching on
the same disk that we booted from, i.e. ubuntu-boot in run mode, and decrypt it
using the partuuid which will be unique instead of the label.

If we are not encrypted, we can just mount ubuntu-data using the helper
directly.

Also rename the decrypt/unlock helper to maybeUnlockAndMount, and print the
requested mounts from the helper directly.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>