tests/session-tool: kill cron session, if any
Despite stopping cron*.service, cron may still be running in a session
when we commence or finish the test. While I have no insight into what
is starting it, we can kill it to avoid any background activity.

This was observed with a session such as:

    Id=2
    User=0
    Name=root
    Timestamp=Wed 2020-04-08 10:01:01 UTC
    TimestampMonotonic=90474701
    VTNr=0
    Remote=no
    Service=crond
    Scope=session-2.scope
    Leader=1659
    Audit=2
    Type=unspecified
    Class=background
    Active=yes
    State=active
    IdleHint=no
    IdleSinceHint=0
    IdleSinceHintMonotonic=0
    LockedHint=no

The session was started at close to 10AM, so perhaps there's some other
source of temporary activities?

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>