wrappers: start after snapd.apparmor.service
We've realized that there's no synchronization between apparmor.service
and any of the service files made by snapd on behalf of snap
applications.

To counter that we could either depend on apparmor.service, which works
correctly on Ubuntu, or on the snapd.apparmor.service, that we ship and
use on non-Ubuntu, apparmor enabled systems. This patch picks the latter
approach.

The reason that this particular service exists is because
apparmor.service normally doesn't consider
/var/lib/snapd/apparmor/profiles and is non-extensible, at least until
the 3.x release is made. On Ubuntu apparmor userspace is patched to
handle /var/lib/snapd/apparmor/profiles.

Interestingly and importantly, specifying a non-existing service in
After is harmless. This is relevant in the case where snapd re-executes
and the snapd.apparmor.service does not exist on the system.

Fixes: https://bugs.launchpad.net/apparmor/+bug/1871148
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>