tests: rewrite timeserver-control test
The test was somewhat flaky and was constantly failing on master.
There are several changes here:

1) The test now handles changing to systemd-timesyncd from either ntp
(on Debian) or from chrony (on Ubuntu 19.10 only). This also enables the
test on work and pass on Debian.

2) The means of querying timesyncd was changed to busctl. This abstracts
away the problem of what the current provider is (those vary across
Ubuntu 14.04 to 20.04) and what specific version of timedatectl can do.

3) The test no longer measures NTPSynchronized, instead it measures NTP,
the actual setting we allow to set via the timeserver-control interface.
This also detaches the test from the racy process of establishing NTP
synchronization which we don't care about.

4) The restore logic now follows the defer-like style to avoid repeating
conditionals and being somewhat easier to follow.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>