cmd/snap-bootstrap/partition/sfdisk: fallback to using label as structure name
This happens for example on an MBR disk, where we don't have partition names it
seems, and only have partition labels. In these situations, we should use the
partition label as the name.

Also upgrade type "c" as reported by sfdisk to "0C" which is what is expected by
the gadget.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>