cmd/snap-bootstrap/partition/lsblk: work with encrypted partitions too
This removes the Equal method of Disk, and instead reimplements that logic with
MountPointIsFromDisk(). It also updates this method and DiskFromMountPoint to
take an Options, which currently just supports specifying that the mount point
is from an decrypted device. If the mount point is a decrypted device, the logic
is different, and instead we need to look for the encrypted physical partition
that provides the decrypted device that is the mount point, and match the
physical partition with the disk, rather than the decrypted device. This is
because the decrypted device mount point will be a mapper device that does not
actually correspond to an actual physical disk.

To implement this, we need to add some additional logic around sysfs and udev,
namely we implement the following steps to find the physical partition that a
decrypted device comes from:
1. read the udev properties for the mountpoint's source path (which will be a dm
   mapper device if the mount point is indeed for a decrypted partition)
2. verify that the device is a dm device by looking at some udev env vars
3. use DM_UUID to get the uuid of the physical backing encrypted partition
4. with the uuid of the of physical backing encrypted partition, get the DEVNAME
   and save that as the path for the rest of the function to search for.

This does not yet implement the necessary mocking to support verifying decrypted
partitions match a disk.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>