cmd/snap-bootstrap/initramfs-mounts: re-order steps, update cross-checks
This changes the steps since we now block and wait for ubuntu-boot, ubuntu-seed,
and then ubuntu-data to be mounted before continuing on to mount the snaps. This
is needed to do sufficiently strong cross-checking of mounted partitions.

This also updates the encrypted cross-check to specify that ubuntu-data may be
a decrypted device, and so we need to hunt down the encrypted physical partition
from that device to perform the required cross-check.

This also fixes a bug in the unlocking code where we were requesting the wrong
partition to be mounted, for systemd to do the right thing, we need to request
mounting the device by label that we told systemd-cryptsetup to decrypt.

Also add an additional check we call systemd-cryptsetup with the right arguments
and re-order some of the tests to match the actual step numbers.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>