github: partition the github action workflows
This commit partitions the spread systems into multiple actions
that are each unit tested with a different go version. This
allows independent restarts and also earlier failures if a
specific go version does break the tests.

The common code in unit-tests-* should be shared somehow,
maybe via a custom action. But that can be done in a followup.