cmd/snap-bootstrap: switch to a 64-byte key for unlocking
AES-256 with XTS block cipher uses 2 keys, so the key size is
512-bits. As the derived key is the same size as this, and we
require the derived key and unlocking key to have the same
entropy, increase the size of the unlocking key to 512-bits too.

As the size of the unlocking key is hardcoded in snap-bootstrap,
keep everything in sync by specifying the cipher and key size
when creating the LUKS2 container.