secboot: fix comment for tpm provisioning function
Contrary to what was stated, this doesn't clear the TPM - it performs
all actions without clearing it first. As a consequence, this will fail
on a device that has previously been provisioned because the current
lockout hierarchy authorization value for the TPM will be incorrect
(it's unknown at this point).

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>