github: run all spread systems in a single go with cached results
This is based on the excellent work from Maciej in PR#8502. This
will run both the 1.9 and latest unit tests and then run all
spread tests. The results of a successful spread run will be
cached so a subsequent re-run will not trigger a spread run
but instead just skip it.

This should make re-running of PRs more feasible.

Unfortunately this requires that the upstream bug
https://github.com/actions/cache/issues/208 is fixed first .(