Moving uc20 tests to new uc20 nested suite
I created a new suite for nested execution just for uc20. Tests on core
suite will be executed on uc16 and uc18, and tests on core20 will be
executed on uc20.

Once the issues related to random reboots are gone/fixed, both test
suites could be merged and the defaults for secure boot and tpm updated.