asserts/internal: introduce Grouping and Groupings
Motivation

In the context of assertion bulk refresh the API supports
dividing the requested assertions in groups, it is natural for
example to have one group for the relevant assertions for each
snap. Some assertions though will then belong to more than one
group, for example store account-keys or developer account
assertions for developers of many snaps on the system.  We want
to detect such cases to request them separately and only once for
better caching properties of the requests. At the low-level
Grouping will help with that with one associated with each
in-play assertion to track to which groups it belongs while
setting up bulk refresh requests. A higher level abstraction
asserts.Pool will leverage this and implement other related
details and features for bulk refresh.