asserts: introduce Pool
Motivation

With the new assertion bulk refresh API we can fetch or update
many assertions together, either assertions related/dependent or
unrelated/independent of each other.

To help driving and managing that this introduces asserts.Pool:

* a Pool can track unresolved assertions to be fetched
* or assertions to be updated
* it can record resolved assertions as fetched from the store and
  implicitly turn unresolved prerequisites into assertions to
  resolve

* assertions can be organized in groups, as the doc comment explains:
    Failure can be tracked isolated to groups, conversely any
    error related to a single group alone will stop any work to
    resolve it.  Independent assertions should not be grouped.

* assertions that are part of or are required by more than one
  group are tracked only once, for example store account-keys or
  developer account assertions for developers of many snaps on
  the system.

The latter is achieved by labeling each assertion with a Grouping
which represents all the groups it belongs to, at the low-level
this uses internal.Groupings/Grouping.

This organization into groupings matches well with the capabilities
of the API and means assertions will be typically requested only
once and grouped in ways that should be cache friendly.