cmd/snap: ignore failures to connect session bus
There are two instances where snap run attempts to connect to session
bus, one when snap run decides to wake up the XDG document portal, the
other is where snap run decides to place the started program in a new
transient unit, a new scope.

The location of the session bus can be provided in one of several ways,
mainly with an environment variable $DBUS_SESSION_BUS_ADDRESS. Another
option is to just try to connect to /run/user/UID/bus.

When using "sudo" to switch from one non-root user to another non-root
user, unless sudo is used with -l, to create a new session, there will
be no bus to connect to.

While not great, tracking is secondary to executing. As such ignore all
errors from establishing connection to session bus.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>