cmd/snap: handle session bus failure better
Work on the test exercising various failures of application tracking
feature has shown me that in some cases we can establish a connection to
the session bus but fail to invoke a DBus method to systemd --user. This
is especially true on Ubuntu 16.04, where systemd --user is just not
available.

This patch changes handling various errors so that failure to connect to
the session bus and failure to talk to systemd --user are handled the
same way. When the invoking user is root we fall back to using the
system bus.

Doing this has also allowed changing some of the error handling so that
failure to track an application is a non-fatal error that logged when
debugging is enabled.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>