cmd/snap: handle NameHasNoOwner
On Ubuntu 14.04 systemd is not running as a session service, since we no
longer try to bus-activate it we now, quickly get a response from the
bus manager that the name org.freedesktop.systemd1 has no owner.

Treat this exactly the same as we handle other cases of old systemd, by
not creating the transient scope.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>