tests: show how app tracking can fail
The interaction of cgroup-based application tracking and systemd version
is sadly complex. This test shows how the interaction looks like across
versions of the stack.

This patch also undoes the no-activate flag, as this is required to
interact with systemd --user on 16.04+.

This patch also contains a fix for dbus-launch leaking when there is no
session bus and go-dbus tries to make one.

The test expands to all core and classic Ubuntu systems.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>