devicestate: do not use snap-boostrap in devicestate to install
This commit changes the way that devicestate creates the
partitions. Instead of calling the snap-bootstrap binary
directly it will import the "bootstrap" package and run it
from inside the code.

This is the minimal PR, in a followup we can remove the
`snap-boostrap create-partitions` command. But this requires
that the spread tests are updated first.