bootloader/ubootenv: support text format
The uboot text format is essentially the same as the native uboot format, but
without the CRC header, any padding and delimits env vars with '\n' instead of
'\0'. This requires us to specify the format when we read the uboot env file in
a few places, but for right now we will continue to use just the native format
everywhere.

This also adds an option to OpenWithFlags to ignore comments, mostly useful for
tests/prototyping.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>