overlord/snapstate: inhibit startup while unlinked
The complete the backend and frontend interaction of
refresh-app-awareness, snapd will write a startup inhibition hint file
when unlinking a snap and remove it when linking a snap.

During the window when a snap is not linked, no new snap application
processes, belonging to a given snap, can be started using "snap run".
Existing snap application processes can fork and create additional
processes but those are already accounted for by the so-called-hard
refresh checker.

While a snap startup is inhibited, snap applications will pause and wait
during "snap run", presenting a user interface appropriate for the
situation at hand.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>