features: add HiddenSnapFolder feature flag
This new feature flag will control the location of the per-user snap
data root. When the flag is unset, the ~/snap location is used. When
the flag is set the new ~/.snapdata location will be used as the root of
all other locations (such as SNAP_USER_DATA and SNAP_USER_COMMON).

The flag is implemented on the go and C sides, as both need to observe
it.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>