tests/main/lxd: add test for snaps inside nested lxd containers not working
Specifically, apparmor does not allow more than one nested namespace, so the
first LXD container works for snapd to run in, but nesting another namespace by
installing snapd inside the nested container should not currently work, so this
test ensures that it doesn't.

Breaking the snapd setup in the container into it's own script makes it easier
to manage this and reduces the duplication.

Also now that 20.04 is released, stop using 19.10 as the 20.04 container.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>