tests: work around journald bug in core16
Restarting or reconfiguring journald on the version of systemd used in
core16 sends SIGPIPE to *all* services, as the journal is closing all
the pipes used for logging.

While this is disastrous for other reasons (all services are affected,
no good solution as outlined in the comment in the test) we don't want
to see random test failures that plague master.

This patch puts a retry loop around the journal manipulation test logic.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>