tests: start pulse to ensure auth cookie is in place
This test has one more race. The first invocation of paplay starts the
pulseaudio.service in the context of the user session, this writes the
(previously absent) pulse authorization cookie into
~/.config/pulse/cookie. The problem is that paplay is racing to read
that same cookie.

To avoid the race, start pulse once and make sure the cookie is present
before invoking paplay for the first time.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>