o/devicestate: support doing system action reboots from recover mode
We want to be able to both re-install and go back to normal run mode when we are
in recover mode, so we need some logic about the current system and the
requested system. Specifically, we support going from run and recover to any
other mode, but only support going from install to install (which is a no-op iff
the system matches the currently installing system).

Also update tests in daemon and devicestate for these new scenarios.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>