o/devicestate: support doing system action reboots from recover mode (#8635)
* tests/o/devicestate: add helper to set device mgr system mode

And use this to set the default system mode as run for tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* o/devicestate: support doing system action reboots from recover mode

We want to be able to both re-install and go back to normal run mode when we are
in recover mode, so we need some logic about the current system and the
requested system. Specifically, we support going from run and recover to any
other mode, but only support going from install to install (which is a no-op iff
the system matches the currently installing system).

Also update tests in daemon and devicestate for these new scenarios.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* tests/o/devicestate: eliminate SetManagerSystemMode, use SetSystemMode

This functionality already existed...

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* tests/o/devicestate: fix test name typo

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* daemon/api_systems_test.go: eliminate ineff assign

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* o/devicestate: catch other invalid system modes

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* o/devicestate: also check the systems match for the no-op same mode case

Currently this is redundant, since systemFromSeed will not set sysAction.Mode to
recover or run if the requested label is not the current system, but this could
hypothetically change at some point.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* o/devicestate: always error for system actions in install mode, extend unit tests

Requesting a system action while in "install" mode is unsupported. Extend unit
tests to cover scenario when the system is not yet seeded.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* daemon: update systems API unit tests

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

Co-authored-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>