snap-mgmt: perform cleanup of user services
Since the introduction of snapd.session-agent.service and the associated
session-level user services there is some more cleanup required when
snapd is being removed.

This patch changes both the Debian postrm script as well as the generic
snap-mgmt command to handle removal of user services as well as the
snapd.session-agent.{service,socket} that snapd creates in the case of
re-execution.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>